feat: minor refactoring to split up code
